make check works, simple reward pool size of 10% of total staked

+++ common/primitives/src/node.rs
@@ -37,7 +37,7 @@ pub type Signature = MultiSignature;
 pub type Index = u32;
 
 /// the time period in blocks that Staking Rewards are based upon
-pub type RewardEra = u32;
+pub type RewardEra = u64;
 
 /// A hash of some data used by the chain.
 pub type Hash = sp_core::H256;
